move dasm sets to their own package
