+++ src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/common/server/level/MixinChunkHolder_Forge.java
@@ -3,7 +3,7 @@ package io.github.opencubicchunks.cubicchunks.mixin.core.common.server.level;
 import io.github.notstirred.dasm.api.annotations.redirect.redirects.AddFieldToSets;
 import io.github.notstirred.dasm.api.annotations.selector.FieldSig;
 import io.github.notstirred.dasm.api.annotations.selector.Ref;
-import io.github.opencubicchunks.cubicchunks.mixin.ForgeSet;
+import io.github.opencubicchunks.cubicchunks.mixin.dasmsets.ForgeSet;
 import io.github.opencubicchunks.cubicchunks.world.level.chunklike.LevelClo;
 import net.minecraft.server.level.ChunkHolder;
 import net.minecraft.world.level.chunk.LevelChunk;
