+++ src/main/java/io/github/opencubicchunks/cubicchunks/mixin/dasmsets/SectionPosToChunkSet.java
@@ -1,4 +1,4 @@
-package io.github.opencubicchunks.cubicchunks.mixin;
+package io.github.opencubicchunks.cubicchunks.mixin.dasmsets;
 
 import io.github.notstirred.dasm.api.annotations.redirect.redirects.MethodRedirect;
 import io.github.notstirred.dasm.api.annotations.redirect.sets.IntraOwnerContainer;
@@ -9,13 +9,13 @@ import io.github.opencubicchunks.cubicchunks.world.level.chunklike.CloPos;
 import net.minecraft.core.SectionPos;
 
 /**
- * Replaces {@link SectionPos}.chunk() in dasm-copied code in cases where a cube position should be returned, rather than a chunk position.
+ * Replaces {@link SectionPos}.chunk() in dasm-copied code in cases where a chunk position should still be returned, not a cube position.
  */
 @RedirectSet
-public interface SectionPosToCubeSet {
+public interface SectionPosToChunkSet {
     @IntraOwnerContainer(owner = @Ref(SectionPos.class))
     abstract class SectionPos_redirects {
         @MethodRedirect(@MethodSig("chunk()Lnet/minecraft/world/level/ChunkPos;"))
-        public native CloPos cc_cube();
+        public native CloPos cc_chunk();
     }
 }
