+++ src/main/java/io/github/opencubicchunks/cubicchunks/mixin/dasmsets/package-info.java
@@ -1,0 +1,7 @@
+@ParametersAreNonnullByDefault
+@MethodsReturnNonnullByDefault
+package io.github.opencubicchunks.cubicchunks.mixin.dasmsets;
+
+import javax.annotation.ParametersAreNonnullByDefault;
+
+import io.github.opencubicchunks.cc_core.annotation.MethodsReturnNonnullByDefault;
